Use new approach for extension usage and tests
