+++ package.json
@@ -31,7 +31,7 @@
     "*"
   ],
   "icon": "icon.png",
-  "main": "./out/src/extension",
+  "main": "./out/extension",
   "contributes": {
     "configuration": {
       "type": "object",
@@ -229,19 +229,23 @@
     ]
   },
   "scripts": {
-    "vscode:prepublish": "tsc -p ./",
-    "compile": "tsc --watch -p ./",
-    "postinstall": "node ./node_modules/vscode/bin/install",
-    "test": "cd testEslint && npm i && cd ../testTslint && npm i && cd .. && cross-env CODE_TESTS_WORKSPACE=testWorkspace.code-workspace node ./node_modules/vscode/bin/test",
+    "vscode:prepublish": "npm run compile",
+    "compile": "tsc -p ./",
+    "watch": "tsc --watch -p ./",
+    "pretest": "npm run compile && cd testEslint && npm i && cd ../testTslint && npm i",
+    "test": "cross-env CODE_TESTS_WORKSPACE=testWorkspace.code-workspace node ./out/test/runTest.js",
     "version": "node ./scripts/version.js && git add CHANGELOG.md"
   },
   "devDependencies": {
+    "@types/glob": "^7.1.1",
     "@types/mocha": "^5.2.7",
     "@types/node": "^12.0.7",
-    "@types/resolve": "0.0.8",
+    "@types/resolve": "^0.0.8",
+    "@types/vscode": "^1.17.0",
     "cross-env": "^5.2.0",
+    "glob": "^7.1.4",
     "mocha": "^6.1.4",
-    "vscode": "^1.1.34"
+    "vscode-test": "^0.4.3"
   },
   "dependencies": {
     "ignore": "^5.1.2",
