+++ src/test/suite/index.ts
@@ -1,0 +1,29 @@
+import * as path from 'path';
+import * as Mocha from 'mocha';
+import * as glob from 'glob';
+
+export function run(testsRoot: string, cb: (error: any, failures?: number) => void): void {
+    // Create the mocha test
+    const mocha = new Mocha({
+        ui: 'tdd',
+    });
+
+    // Use any mocha API
+    mocha.useColors(true);
+
+    glob('**/**.test.js', { cwd: testsRoot }, (err, files) => {
+        if (err) {
+            return cb(err);
+        }
+
+        // Add files to the test suite
+        files.forEach(f => mocha.addFile(path.resolve(testsRoot, f)));
+
+        try {
+            // Run the mocha test
+            mocha.run(failures => cb(null, failures));
+        } catch (err) {
+            cb(err);
+        }
+    });
+}
